<Persistence, Feature> Save Regions, Explore Menu
Regions are now saved.
Removed MapGenerator- now part of RegionManager.
Created Explore Menu scripts and prefab.

+++ Night Unity Files/Assets/Scripts/Game/Characters/Attributes/SurvivalAttributes.cs
@@ -81,8 +81,7 @@ namespace Game.Characters.Attributes
             SaveAttribute(doc, nameof(Dehydration), Dehydration);
             SaveAttribute(doc, nameof(Hunger), Hunger);
             SaveAttribute(doc, nameof(Thirst), Thirst);
-
-            SaveController.CreateNodeAndAppend("Weight", doc, (int) Weight);
+            SaveController.CreateNodeAndAppend("Weight", doc, Weight);
         }
         
         private void Drink()
